+++ src/collective/elastic/ingest/ingest.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from .analysis import update_analysis
 from .elastic import get_ingest_client
 from .logging import logger
 from .mapping import create_or_update_mapping
@@ -49,6 +50,9 @@ def ingest(content, full_schema, index_name):
     # preprocess content and schema
     preprocess(content, full_schema)
     if full_schema:
+        # first update_analysis, then create_or_update_mapping:
+        # mapping can use analyzers from analysis.json
+        update_analysis(index_name)
         create_or_update_mapping(full_schema, index_name)
         if not STATES["pipelines_created"]:
             setup_ingest_pipelines(full_schema, index_name)
@@ -56,7 +60,7 @@ def ingest(content, full_schema, index_name):
     info = {"expansion_fields": EXPANSION_FIELDS}
     postprocess(content, info)
 
-    logger.debug(pformat(content))
+    logger.info(f"Index content: {pformat(content)}")
     es = get_ingest_client()
     es_kwargs = dict(
         index=index_name,