+++ src/collective/elastic/ingest/preprocessing_test.py
@@ -9,7 +9,7 @@ import pytest
 
 
 def test_find_last_container_in_path_flat():
-    root = {'corona': 'lockdown'}
+    root = {"corona": "lockdown"}
     container, key = preprocessing._find_last_container_in_path(root, ["corona"])
     assert container is root
     assert key == "corona"
@@ -17,10 +17,13 @@ def test_find_last_container_in_path_flat():
 
 def test_find_last_container_in_path_nested():
     date_container = {"date": "2020-03-16"}
-    root = {'corona': {'lockdown': date_container}}
-    container, key = preprocessing._find_last_container_in_path(root, ["corona", "lockdown", "date"])
+    root = {"corona": {"lockdown": date_container}}
+    container, key = preprocessing._find_last_container_in_path(
+        root, ["corona", "lockdown", "date"]
+    )
     assert container is date_container
     assert key == "date"
+
 
 # ------------------------------------------------------------------------------
 # actions
@@ -33,30 +36,22 @@ def test_action_rewrite():
             "rid": ridvalue,
         }
     }
-    config = {
-        "source": "@components/rid",
-        "target": "rid"
-    }
+    config = {"source": "@components/rid", "target": "rid"}
     preprocessing.action_rewrite(root, {}, config)
     assert "rid" in root
     assert root["rid"] == ridvalue
-    assert "rid" not in root['@components']
+    assert "rid" not in root["@components"]
+
 
 def test_action_rewrite_non_existing():
-    root = {
-        "@components": {}
-    }
-    config = {
-        "source": "@components/rid",
-        "target": "rid"
-    }
+    root = {"@components": {}}
+    config = {"source": "@components/rid", "target": "rid"}
     preprocessing.action_rewrite(root, {}, config)
     assert "rid" not in root
 
+
 def test_action_rewrite_non_existing_forced():
-    root = {
-        "@components": {}
-    }
+    root = {"@components": {}}
     config = {
         "source": "@components/rid",
         "target": "rid",