+++ CHANGES.rst
@@ -5,5 +5,11 @@ Changelog
 1.0a1 (unreleased)
 ------------------
 
+- Add optional configuration of text analysis (stemmer, decompunder, etc)
+  [ksuess]
+
+- Keep source on rewrite 
+  [ksuess]
+
 - Initial release.
   [jensens]